requirements.txt
requests
PyJWT
gen3
boto3
botocore
jmespath
s3transfer

